requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
faster-whisper==1.1.1
supabase==2.11.0
httpx==0.28.1
